signal: Move signal to new package signal
